[RF] Convert RooSimultaneous::_pdfProxyList from TList to std::map.
RooSimultaneous was using a TList of RooRealProxies to keep track of
PDFs for each category. This always required casting and makes
retrieving a certain PDF slow, because they have to be found by name.

After modernising proxies and categories, PDFs can now be stored in a
map with key=category index.
A schema evolution rule ensures that PDFs are read correctly from old
files, and when a PDF is looked up by index, the mapping of index
names to PDFs is rebuilt using the category names+index numbers.