[RF] Rewrite RooIntegrator1D.
- Re-implement trapezoid and midpoint integration rules using fast batch
evaluations. Also re-implement Romberg-style series acceleration.
- Reduce class members significantly.

Note:
The old integrator had a buggy check for convergence, such that it was
often not reaching the desired precision. The new integrator will
therefore produce different values in a few tests, but the unit tests
to follow this commit will show that the new integrator is usually
right.