[RF] Get rid of a dynamic_cast in RooDataHist.
Since categories and real-valued objects are used as observables in RooDataHist,
cross-casting to RooAbsLValue is required in a few places. Since this
can only be done with a dynamic_cast, the casts are executed once during
initialisation, and results are used during fitting.
Here, the last three fit-time dynamic_casts are removed.