[TODO][CMake] Make XRootD config target-based.
Create an XRootD target for usage in ROOT
ROOT::XRootD

When xrootd is found in system directories, the target will be
an IMPORTED target, whereas it's a normal library when XRootD is a builtin.
This also solves the problem of ROOT's build system picking up ROOT
headers when it is trying to include xrootd.

All packages inside ROOT should prefer this target over using any variables.
This way, an update of XRootD's CMake will only affect one single location in ROOT.

Also refactor builtin XRootD. Synchronise the variables that builtin vs
SearchInstalled are setting, and use those to configure the ROOT-internal
target.