Fix hardware_destructive_interference_size across TUs.
Determine the interference size only once in the beginning with the compiler
settings used ROOT when is configured, and keep this value afterwards.

Since the value of std::hardware_destructive_interference_size can vary
with compiler versions, standard library, or tune settings, it might
change across translation units in ROOT, breaking the ABI of RDF/RVec.
For this reason, gcc warns about using it in any header and/or public
interface.
When e.g. using address sanitizer, JITted code was trying to use the
standard C++ value, whereas it was undefined in precompiled code due to
how asan works.
Similarly, when users use an installed ROOT, they might include a header
that uses the C++ interference size, which might differ from the one that
ROOT was compiled with.

For these reasons. ROOT is now configured with one global value to keep
the ABI stable, even if this value might not be optimal when this ROOT is
later run on a different architecture.