[roottest] Fix output file path in nbdiff.py
nbconvert was creating an invalid path when creating an output notebook.
This seems to result from a behaviour change that doesn't anymore trim a
redundant path from a notebook name ("a/b/notebook" was transformed to
"notebook" if the input notebook is in a/b).
In nbconvert 7.16.6, this became "a/b/a/b/notebook".

Since nbconvert 5.0, the notebook is anyway placed next to the input
notebook instead of the current working directory, so the output path
can be removed from the command.